refactor: delete-todo 분리
