+++ src/features/delete-todo/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useDeleteTodo';
