+++ src/features/delete-todo/model/hooks/useDeleteTodo.ts
@@ -1,0 +1,13 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { useTodoCustomMutation } from '@/shared/lib/utils/useTodoCustomMutation';
+import { deleteTodo } from '../../api';
+
+interface DeleteTodoMutationParams {
+  todolistId: string;
+  todoId: string;
+}
+export const useDeleteTodoMutation = () =>
+  useTodoCustomMutation<DeleteTodoMutationParams, any>(
+    ({ todolistId, todoId }) => deleteTodo(todolistId, todoId),
+    ['todolist'],
+  );
