+++ src/features/todo/api/todolist.ts
@@ -27,10 +27,3 @@ export const updateTodoOrder = async (
 
   return parsedResponse;
 };
-
-export const deleteTodo = async (todolistId: string, todoId: string) => {
-  const endpoint = `/api/todolist/${todolistId}/todo/${todoId}`;
-  const parsedResponse = await clientFetch.delete(endpoint);
-
-  return parsedResponse;
-};
